Code for the timetable creation. Need to fix the constraints.

+++ Codigo/TimetableGenerator/src/main/java/thesis/model/persistence/entities/ConfigurationEntity.java
@@ -11,7 +11,7 @@ public class ConfigurationEntity {
     private int id;
 
     @Column(name = "number_days")
-    private int numberDays;
+    private short numberDays;
 
     @Column(name = "number_weeks")
     private int numberWeeks;
@@ -24,7 +24,7 @@ public class ConfigurationEntity {
 
     public ConfigurationEntity() {}
 
-    public ConfigurationEntity(int numberDays, int numberWeeks, int slotsPerDay) {
+    public ConfigurationEntity(short numberDays, int numberWeeks, int slotsPerDay) {
         this.id = 0;
         this.numberDays = numberDays;
         this.numberWeeks = numberWeeks;
@@ -35,11 +35,11 @@ public class ConfigurationEntity {
         return id;
     }
 
-    public int getNumberDays() {
+    public short getNumberDays() {
         return numberDays;
     }
 
-    public void setNumberDays(int numberDays) {
+    public void setNumberDays(short numberDays) {
         this.numberDays = numberDays;
     }
